Move useProjectContext into entities/project
Project domain state (projectId, statuses, transitions, doneStatusIds) lived in
the domain-agnostic shared/model layer. Move the context and its hook to
entities/project, where they belong — every consumer is a feature or widget,
which may import entities, so no boundary is crossed. Consumers now import it
alongside the other project-domain helpers they already pull from
@/entities/project. Tests that stubbed it re-point their mock accordingly.
