+++ apps/hobom-system/src/features/issue-detail/ui/IssueMetaSection.tsx
@@ -2,10 +2,10 @@ import { useCallback, useMemo, useState } from "react";
 import { AddOutlined } from "hobom-design-system/icons";
 import { useQuery } from "hobom-data";
 import * as stylex from "@stylexjs/stylex";
-import { useProjectContext, usePopoverState } from "@/shared/model";
+import { usePopoverState } from "@/shared/model";
 import { ISSUE_KIND_LABEL, ISSUE_PRIORITY_LABEL } from "@/entities/issue";
 import { ISSUE_PRIORITY_REGISTRY, ParentIssueAutocomplete } from "@/entities/issue/ui";
-import { getStatusName, getStatusColor } from "@/entities/project";
+import { getStatusName, getStatusColor, useProjectContext } from "@/entities/project";
 import { projectLabelQueries } from "@/entities/project-label";
 import { ProjectLabelPicker } from "@/entities/project-label/ui";
 import { Hb } from "@/shared/ui";
